+++ ReportsWebApiLayer/Controllers/TasksController.cs
@@ -6,6 +6,7 @@ using ReportsLibrary.Enums;
 using ReportsLibrary.Tasks;
 using ReportsLibrary.Tasks.TaskChangeCommands;
 using ReportsWebApiLayer.DataTransferObjects;
+using ReportsWebApiLayer.Extensions;
 
 namespace ReportsWebApiLayer.Controllers;
 
@@ -32,8 +33,12 @@ public class TasksController : ControllerBase
     [HttpGet("{taskId}", Name = "GetTaskById")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
+    [ProducesResponseType(typeof(List<string>), StatusCodes.Status400BadRequest)]
     public async Task<ActionResult<FullReportsTaskDto>> GetTask([FromRoute] Guid taskId)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState.GetErrorMessages());
+
         try
         {
             ReportsTask reportsTask = await _tasksService.GetTaskById(taskId);
@@ -47,11 +52,15 @@ public class TasksController : ControllerBase
 
     [HttpPost]
     [ProducesResponseType(StatusCodes.Status201Created)]
-    public async Task<IActionResult> CreateTask(string taskName, Guid creatorId)
+    [ProducesResponseType(typeof(List<string>), StatusCodes.Status400BadRequest)]
+    public async Task<IActionResult> CreateTask(string taskName)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState.GetErrorMessages());
+
         try
         {
-            ReportsTask reportsTask = await _tasksService.CreateTask(taskName, creatorId);
+            ReportsTask reportsTask = await _tasksService.CreateTask(taskName);
 
             return CreatedAtRoute(
                 "GetTaskById", new { taskId = reportsTask.Id }, _mapper.Map<ReportsTaskDto>(reportsTask));
@@ -65,9 +74,13 @@ public class TasksController : ControllerBase
     // GET: Tasks/1
     [HttpGet("byCreationTime/{creationTime}", Name = "GetTasksByCreationTime")]
     [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(List<string>), StatusCodes.Status400BadRequest)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<ActionResult<List<FullReportsTaskDto>>> FindTasksByCreationTime([FromRoute] DateTime creationTime)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState.GetErrorMessages());
+
         IReadOnlyCollection<ReportsTask> tasksByCreationTime = await _tasksService.FindTasksByCreationTime(creationTime);
 
         if (tasksByCreationTime == null || tasksByCreationTime.Count == 0) return NotFound();
@@ -78,10 +91,14 @@ public class TasksController : ControllerBase
     // GET: Tasks/1
     [HttpGet("byModificationTime/{modificationTime}", Name = "GetTasksByModificationTime")]
     [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(List<string>), StatusCodes.Status400BadRequest)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<ActionResult<List<FullReportsTaskDto>>> FindTasksByModificationTime(
         [FromRoute] DateTime modificationTime)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState.GetErrorMessages());
+
         IReadOnlyCollection<ReportsTask> tasksByModificationTime =
             await _tasksService.FindTasksByModificationDate(modificationTime);
 
@@ -93,9 +110,13 @@ public class TasksController : ControllerBase
     // GET: Tasks/1
     [HttpGet("byEmployee/{employeeId}", Name = "GetTasksByEmployee")]
     [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(List<string>), StatusCodes.Status400BadRequest)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<ActionResult<List<FullReportsTaskDto>>> GetTasksByEmployee([FromRoute] Guid employeeId)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState.GetErrorMessages());
+
         IReadOnlyCollection<ReportsTask> employeeTasks = await _tasksService.FindTasksByEmployeeId(employeeId);
 
         if (employeeTasks == null || employeeTasks.Count == 0) return NotFound();
@@ -106,9 +127,13 @@ public class TasksController : ControllerBase
     // GET: Tasks/1
     [HttpGet("ModifiedByEmployee/{employeeId}", Name = "GetTasksModifiedByEmployee")]
     [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(List<string>), StatusCodes.Status400BadRequest)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<ActionResult<List<FullReportsTaskDto>>> GetTasksModifiedByEmployee([FromRoute] Guid employeeId)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState.GetErrorMessages());
+
         IReadOnlyCollection<ReportsTask> tasksModifiedByEmployee
             = await _tasksService.FindsTaskModifiedByEmployeeId(employeeId);
 
@@ -120,10 +145,14 @@ public class TasksController : ControllerBase
     // GET: Tasks/1
     [HttpGet("CreatedByEmployeeSubordinates/{employeeId}", Name = "GetTasksCreatedByEmployeeSubordinates")]
     [ProducesResponseType(StatusCodes.Status200OK)]
+    [ProducesResponseType(typeof(List<string>), StatusCodes.Status400BadRequest)]
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<ActionResult<List<FullReportsTaskDto>>> GetTasksCreatedByEmployeeSubordinates(
         [FromRoute] Guid employeeId)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState.GetErrorMessages());
+
         IReadOnlyCollection<ReportsTask> tasksCreatedByEmployeeSubordinates
             = await _tasksService.FindTasksCreatedByEmployeeSubordinates(employeeId);
 
@@ -135,103 +164,127 @@ public class TasksController : ControllerBase
     [HttpPut("{taskId}/owner")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
-    public Task<IActionResult> SetOwner(
+    [ProducesResponseType(typeof(List<string>), StatusCodes.Status400BadRequest)]
+    public async Task<IActionResult> SetOwner(
         [Required] Guid taskId,
         [Required] Guid changerId,
         [FromQuery] Guid ownerId)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState.GetErrorMessages());
+
         try
         {
-            _tasksService.UseChangeTaskCommand(taskId, changerId, new SetTaskOwnerCommand(ownerId));
-            return Task.FromResult<IActionResult>(Ok());
+            await _tasksService.UseChangeTaskCommand(taskId, changerId, new SetTaskOwnerCommand(ownerId));
+            return Ok();
         }
         catch (Exception e)
         {
-            return Task.FromResult<IActionResult>(Problem(e.Message));
+            return Problem(e.Message);
         }
     }
 
     [HttpPut("{taskId}/content")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
-    public Task<IActionResult> SetContent(
+    [ProducesResponseType(typeof(List<string>), StatusCodes.Status400BadRequest)]
+    public async Task<IActionResult> SetContent(
         [Required] Guid taskId,
         [Required] Guid changerId,
         [FromQuery] string content)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState.GetErrorMessages());
+
         try
         {
-            _tasksService.UseChangeTaskCommand(taskId, changerId, new SetTaskContentCommand(content));
-            return Task.FromResult<IActionResult>(Ok());
+            await _tasksService.UseChangeTaskCommand(taskId, changerId, new SetTaskContentCommand(content));
+            return Ok();
         }
         catch (Exception e)
         {
-            return Task.FromResult<IActionResult>(Problem(e.Message));
+            return Problem(e.Message);
         }
     }
 
     [HttpPut("{taskId}/comment")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
-    public Task<IActionResult> AddComment(
+    [ProducesResponseType(typeof(List<string>), StatusCodes.Status400BadRequest)]
+    public async Task<IActionResult> AddComment(
         [Required] Guid taskId,
         [FromQuery] Guid changerId,
         [FromQuery] string comment)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState.GetErrorMessages());
+
         try
         {
-            _tasksService.UseChangeTaskCommand(taskId, changerId, new AddTaskCommentCommand(comment));
-            return Task.FromResult<IActionResult>(Ok());
+            await _tasksService.UseChangeTaskCommand(taskId, changerId, new AddTaskCommentCommand(comment));
+            return Ok();
         }
         catch (Exception e)
         {
-            return Task.FromResult<IActionResult>(Problem(e.Message));
+            return Problem(e.Message);
         }
     }
 
     [HttpPut("{taskId}/title")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
-    public Task<IActionResult> SetTitle(
+    [ProducesResponseType(typeof(List<string>), StatusCodes.Status400BadRequest)]
+    public async Task<IActionResult> SetTitle(
         [Required] Guid taskId,
         [FromQuery] Guid changerId,
         [FromQuery] string title)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState.GetErrorMessages());
+
         try
         {
-            _tasksService.UseChangeTaskCommand(taskId, changerId, new SetTaskTitleCommand(title));
-            return Task.FromResult<IActionResult>(Ok());
+            await _tasksService.UseChangeTaskCommand(taskId, changerId, new SetTaskTitleCommand(title));
+            return Ok();
         }
         catch (Exception e)
         {
-            return Task.FromResult<IActionResult>(Problem(e.Message));
+            return Problem(e.Message);
         }
     }
 
     [HttpPut("{taskId}/state")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
-    public Task<IActionResult> SetState(
+    [ProducesResponseType(typeof(List<string>), StatusCodes.Status400BadRequest)]
+    public async Task<IActionResult> SetState(
         [Required] Guid taskId,
         [FromQuery] Guid changerId,
         [FromQuery] TaskStates state)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState.GetErrorMessages());
+
         try
         {
-            _tasksService.UseChangeTaskCommand(taskId, changerId, new SetTaskStateCommand(state));
-            return Task.FromResult<IActionResult>(Ok());
+            await _tasksService.UseChangeTaskCommand(taskId, changerId, new SetTaskStateCommand(state));
+            return Ok();
         }
         catch (Exception e)
         {
-            return Task.FromResult<IActionResult>(Problem(e.Message));
+            return Problem(e.Message);
         }
     }
 
     [HttpDelete("{taskId}")]
     [ProducesResponseType(StatusCodes.Status200OK)]
     [ProducesResponseType(StatusCodes.Status204NoContent)]
+    [ProducesResponseType(typeof(List<string>), StatusCodes.Status400BadRequest)]
     public IActionResult Delete(Guid taskId)
     {
+        if (!ModelState.IsValid)
+            return BadRequest(ModelState.GetErrorMessages());
+
         try
         {
             _tasksService.RemoveTaskById(taskId);
